and fold static methods into Message
