+++ src/Carnac.Logic/Carnac.Logic.csproj
@@ -91,7 +91,6 @@
     <Compile Include="KeyMonitor\InterceptKeys.cs" />
     <Compile Include="KeyMonitor\KeyDirection.cs" />
     <Compile Include="KeysController.cs" />
-    <Compile Include="MessageMerger.cs" />
     <Compile Include="Models\Processinfo.cs" />
     <Compile Include="ShortcutAccumulator.cs" />
     <Compile Include="KeyPressDefinition.cs" />
